Implement basic unit movement validation
I ended up putting this logic inside the unit group instead of the unit container. The group just needs to reference the group container's all_units array. This way each group can determine its default directions in an exported variable and initialize each unit's corresponding directions variable on initialization. I did this because when a unit goes sicko mode when it reaches the end of the board, it can just add the two directions it doesn't have and the group can check for that when it does its check for available moves in its units. Okay great